requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cssbeautifier==1.15.4
cycler==0.12.1
djlint==1.36.4
EditorConfig==0.17.0
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.0
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.2
Werkzeug==3.1.3
